+++ src/generic/WholeMolecules.cpp
@@ -151,7 +151,7 @@ ActionAtomistic(ao)
       std::vector< std::vector<AtomNumber> > backatoms;
       moldat[0]->getBackbone( resstrings, moltype, backatoms );
       for(unsigned i=0;i<backatoms.size();++i){
-          log.printf("  atoms in entity %d : ", groups.size()+1 );
+          log.printf("  atoms in entity %u : ", static_cast<unsigned>(groups.size()+1));
           for(unsigned j=0;j<backatoms[i].size();++j) log.printf("%d ",backatoms[i][j].serial() );
           log.printf("\n");
           groups.push_back( backatoms[i] );
