+++ src/secondarystructure/SecondaryStructureRMSD.cpp
@@ -140,7 +140,7 @@ void SecondaryStructureRMSD::readBackboneAtoms( const std::string& moltype, std:
 void SecondaryStructureRMSD::addColvar( const std::vector<unsigned>& newatoms ){
   if( colvar_atoms.size()>0 ) plumed_assert( colvar_atoms[0].size()==newatoms.size() );
   if( verbose_output ){
-     log.printf("  Secondary structure segment %d contains atoms : ", colvar_atoms.size()+1);
+     log.printf("  Secondary structure segment %u contains atoms : ", static_cast<unsigned>(colvar_atoms.size()+1));
      for(unsigned i=0;i<newatoms.size();++i) log.printf("%d ",all_atoms(newatoms[i]).serial() );
      log.printf("\n");
   }
